requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==25.3.0
blinker==1.8.2
certifi==2024.8.30
click==8.1.7
deprecation==2.1.0
distro==1.9.0
Flask==3.0.3
Flask-Cors==5.0.0
frozenlist==1.6.0
gotrue==2.12.0
gunicorn==23.0.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.6
httpx==0.27.2
hyperframe==6.1.0
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.6.1
MarkupSafe==3.0.2
multidict==6.4.3
openai==1.52.2
packaging==24.2
pluggy==1.5.0
postgrest==1.0.1
propcache==0.3.1
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.10.1
pytest==8.3.5
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
realtime==2.4.3
six==1.17.0
sniffio==1.3.1
storage3==0.11.3
StrEnum==0.4.15
supabase==2.15.1
supafunc==0.9.4
tqdm==4.66.5
typing_extensions==4.13.2
websockets==14.2
Werkzeug==3.0.4
yarl==1.20.0
